+++ forui/lib/theme.dart
@@ -3,6 +3,7 @@
 library forui.theme;
 
 export 'src/theme/color_scheme.dart';
+export 'src/theme/form_style.dart';
 export 'src/theme/style.dart';
 export 'src/theme/theme.dart';
 export 'src/theme/theme_data.dart';